Implement maximum squares per row validation:
This will fix a bug that was happening on small screens, where the square size was being
rendered too small, generating performance issues.
Now the square size is set to a minimum of 10 pixels.
Also there's a validation so that the total number of squares per row
doesn't overflow the width of the screen (mainly on mobile), when generating
10 pixels square. So besides resizing the squares, the total number of squares per row
is also adjusted according to the screen width, forcing a maximum number of
squares per row to fit the screen properly.