Revert to validating square size to not overflow screen width:
I tried a couple of approaches to dynamically resize the canvas (and the body)
but that has lead to canvas being to big to handle properly on small screens.
So I am now opting to revert to previous behaviour, where the squares will be
resized when screen orientation changes, but never overflowing the screen width.
Although there's still a known issue with that:
when on lanscape mode, using a large number of rows, and changing back to
portrait, the canvas and its inner squares, will be resized to fit the new
screen width, leading to a canvas with tiny squares (that was the issue I was
trying to fix on previous commits). I think that approach is ok, as the user can
toggle back to landscape orientation and the squares will be properly sized again,
and trying to rerender the canvas when toggling back to portrait would lead to user
loosing the coloring already done.